Refactored UserTest and SubmissionTest classes. Fixed mark/unmark NSFW test method
