+++ src/test/java/com/reddit/test/UserTest.java
@@ -1,27 +1,26 @@
 package com.reddit.test;
 
 import static org.junit.Assert.*;
+
+import com.reddit.utils.TestUtils;
 import im.goel.jreddit.user.User;
 
 import org.junit.Test;
 
 
-
+/**
+ * Class for testing User-related methods
+ *
+ * @author Karan Goel
+ * @author Raul Rene Lepsa
+ */
 public class UserTest {
+
 	@Test
 	public void test() {
-		// Don't hax meh plz
-		User user = new User("jRedditTest", "jRedditTest"); // Enter your username and password here!
-		try {
-			user.connect();
-            System.out.println(user.commentKarma());
-		} catch (Exception exception) {
-			exception.printStackTrace();
-		}
+		User user = TestUtils.createAndConnectUser();
 
-
-		assertNotNull(
-                "The user's modhash should never be null, lest no API methods function properly. (Invoke the connect function to remedy this)",
-                user.getModhash());
+        assertNotNull(user);
+		assertNotNull("The user's modhash should never be null", user.getModhash());
 	}
 }
